refactor : Event 클래스의 필드 타입을 변경 및 prevSection 추가
